refactor from macos
