--- bindings/README.md
@@ -1,7 +1,0 @@
-# Rookie bindings
-
-## Python
-
-```console
-pip3 install rookiepy
-```
