--- rookie-rs/examples/chrome/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "chrome"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-
-[dependencies]
-rookie = { path = "../../", version = "*" }
