--- rookie-rs/examples/chrome/src/main.rs
@@ -1,4 +1,0 @@
-fn main() {
-  let cookies = rookie::chrome(None).unwrap();
-  println!("{cookies:?}");
-}
