+++ rookie-rs/examples/simple.rs
