+++ rookie-rs/src/browser/chromium.rs
@@ -1,8 +1,10 @@
 use crate::common::{date, enums::*, sqlite};
 use cfg_if::cfg_if;
 use eyre::{bail, ContextCompat, Result};
-use log::{info, warn};
 use std::path::PathBuf;
+
+#[cfg(target_os = "macos")]
+use crate::macos::secrets;
 
 cfg_if! {
     if #[cfg(target_os = "windows")] {
@@ -12,8 +14,51 @@ cfg_if! {
         use crate::winapi;
         use eyre::Context;
     } else if #[cfg(unix)] {
-        use crate::common::secrets;
+
     }
+}
+
+#[cfg(target_os = "windows")]
+pub fn chromium_based(
+  key: PathBuf,
+  db_path: PathBuf,
+  domains: Option<Vec<&str>>,
+) -> Result<Vec<Cookie>> {
+  // Use DPAPI
+  let content = std::fs::read_to_string(&key)?;
+  let key_dict: serde_json::Value =
+    serde_json::from_str(content.as_str()).context("Can't read json file")?;
+
+  let os_crypt = key_dict.get("os_crypt").context("Can't get os crypt")?;
+
+  let key64 = os_crypt
+    .get("encrypted_key")
+    .context("Can't get encrypted_key")?
+    .as_str()
+    .context("Can't convert encrypted_key to str")?;
+
+  let keys = get_keys(key64)?;
+  query_cookies(keys, db_path, domains)
+}
+
+#[cfg(unix)]
+pub fn chromium_based(
+  config: &BrowserConfig,
+  db_path: PathBuf,
+  domains: Option<Vec<&str>>,
+) -> Result<Vec<Cookie>> {
+  // Simple AES
+  let keys = get_keys(config)?;
+  query_cookies(keys, db_path, domains)
+}
+
+#[cfg(unix)]
+fn create_pbkdf2_key(password: &str, salt: &[u8; 9], iterations: u32) -> Vec<u8> {
+  use pbkdf2::pbkdf2_hmac;
+  use sha1::Sha1;
+  let mut output = [0u8; 16];
+  pbkdf2_hmac::<Sha1>(password.as_bytes(), salt, iterations, &mut output);
+  output.to_vec()
 }
 
 #[cfg(target_os = "windows")]
@@ -26,78 +71,63 @@ fn get_keys(key64: &str) -> Result<Vec<Vec<u8>>> {
   Ok(keys)
 }
 
-#[cfg(unix)]
-fn create_pbkdf2_key(password: &str, salt: &[u8; 9], iterations: u32) -> Vec<u8> {
-  use pbkdf2::pbkdf2_hmac;
-  use sha1::Sha1;
-  let mut output = [0u8; 16];
-  pbkdf2_hmac::<Sha1>(password.as_bytes(), salt, iterations, &mut output);
-  output.to_vec()
-}
-
-#[cfg(unix)]
+#[cfg(target_os = "linux")]
 fn get_keys(config: &BrowserConfig) -> Result<Vec<Vec<u8>>> {
   // AES CBC key
 
   let salt = b"saltysalt";
   let iterations: u32;
-  #[cfg(target_os = "linux")]
-  {
-    iterations = 1;
-  }
 
-  #[cfg(target_os = "macos")]
-  {
-    iterations = 1003;
-  }
+  iterations = 1;
 
   let mut keys: Vec<Vec<u8>> = vec![];
-
-  #[cfg(target_os = "linux")]
-  {
-    if let Ok(passwords) = secrets::get_passwords(config.os_crypt_name.unwrap_or("")) {
-      for password in passwords {
-        let key = create_pbkdf2_key(password.as_str(), salt, iterations);
-        keys.push(key);
-      }
+  if let Ok(passwords) = secrets::get_passwords(config.os_crypt_name.unwrap_or("")) {
+    for password in passwords {
+      let key = create_pbkdf2_key(password.as_str(), salt, iterations);
+      keys.push(key);
     }
-    // default keys
-    let key = create_pbkdf2_key("peanuts", salt, iterations);
-    keys.push(key);
-    let key = create_pbkdf2_key("", salt, iterations);
-    keys.push(key);
   }
-  #[cfg(target_os = "macos")]
-  {
-    let key_service = config.osx_key_service.context("missing osx_key_service")?;
-    let key_user = config.osx_key_user.context("missing osx_key_user")?;
-    let password =
-      secrets::get_osx_keychain_password(key_service, key_user).unwrap_or("peanuts".to_string());
-
-    let key = create_pbkdf2_key(password.as_str(), salt, iterations);
-    keys.push(key);
-
-    let key = create_pbkdf2_key("peanuts", salt, iterations);
-    keys.push(key);
-    let key = create_pbkdf2_key("", salt, iterations);
-    keys.push(key);
-  }
-
-  // keychain key
-
   // default keys
+  let key = create_pbkdf2_key("peanuts", salt, iterations);
+  keys.push(key);
+  let key = create_pbkdf2_key("", salt, iterations);
+  keys.push(key);
 
   Ok(keys)
 }
 
+#[cfg(target_os = "macos")]
+fn get_keys(config: &BrowserConfig) -> Result<Vec<Vec<u8>>> {
+  let salt = b"saltysalt";
+  let iterations: u32;
+
+  iterations = 1003;
+
+  let mut keys: Vec<Vec<u8>> = vec![];
+
+  let key_service = config.osx_key_service.context("missing osx_key_service")?;
+  let key_user = config.osx_key_user.context("missing osx_key_user")?;
+  let password =
+    secrets::get_osx_keychain_password(key_service, key_user).unwrap_or("peanuts".to_string());
+
+  let key = create_pbkdf2_key(password.as_str(), salt, iterations);
+  keys.push(key);
+
+  let key = create_pbkdf2_key("peanuts", salt, iterations);
+  keys.push(key);
+  let key = create_pbkdf2_key("", salt, iterations);
+  keys.push(key);
+
+  Ok(keys)
+}
+
+/// Decrypt cookie value using aes GCM
 #[cfg(target_os = "windows")]
 fn decrypt_encrypted_value(
   value: String,
   encrypted_value: &[u8],
   keys: Vec<Vec<u8>>,
 ) -> Result<String> {
-  // gcm
-
   let key_type = &encrypted_value[..3];
   if !value.is_empty() || !(key_type == b"v11" || key_type == b"v10") {
     // unknown key_type or value isn't encrypted
@@ -122,6 +152,7 @@ fn decrypt_encrypted_value(
   bail!("decrypt_encrypted_value failed")
 }
 
+/// Decrypt cookie value using aes cbc
 #[cfg(unix)]
 fn decrypt_encrypted_value(
   value: String,
@@ -162,7 +193,7 @@ fn decrypt_encrypted_value(
           return Ok(decoded);
         }
         Err(_) => {
-          warn!("Error in decode decrypt value with utf8");
+          log::warn!("Error in decode decrypt value with utf8");
           return Ok("".into());
         }
       }
@@ -176,16 +207,17 @@ fn query_cookies(
   db_path: PathBuf,
   domains: Option<Vec<&str>>,
 ) -> Result<Vec<Cookie>> {
+  /// On windows release file lock
   #[cfg(target_os = "windows")]
   {
     let db_path_str = db_path.to_str().context("Can't convert db path to str")?;
-    warn!("Unlocking Chrome database... This may take a while (sometimes up to a minute)");
+    log::warn!("Unlocking Chrome database... This may take a while (sometimes up to a minute)");
     unsafe {
       winapi::release_file_lock(db_path_str);
     }
   }
 
-  info!(
+  log::info!(
     "Creating SQLite connection to {}",
     db_path.to_str().unwrap_or("")
   );
@@ -240,38 +272,4 @@ fn query_cookies(
     cookies.push(cookie);
   }
   Ok(cookies)
-}
-
-#[cfg(target_os = "windows")]
-pub fn chromium_based(
-  key: PathBuf,
-  db_path: PathBuf,
-  domains: Option<Vec<&str>>,
-) -> Result<Vec<Cookie>> {
-  // Use DPAPI
-  let content = std::fs::read_to_string(&key)?;
-  let key_dict: serde_json::Value =
-    serde_json::from_str(content.as_str()).context("Can't read json file")?;
-
-  let os_crypt = key_dict.get("os_crypt").context("Can't get os crypt")?;
-
-  let key64 = os_crypt
-    .get("encrypted_key")
-    .context("Can't get encrypted_key")?
-    .as_str()
-    .context("Can't convert encrypted_key to str")?;
-
-  let keys = get_keys(key64)?;
-  query_cookies(keys, db_path, domains)
-}
-
-#[cfg(unix)]
-pub fn chromium_based(
-  config: &BrowserConfig,
-  db_path: PathBuf,
-  domains: Option<Vec<&str>>,
-) -> Result<Vec<Cookie>> {
-  // Simple AES
-  let keys = get_keys(config)?;
-  query_cookies(keys, db_path, domains)
 }
