+++ rookie-rs/src/common/date.rs
@@ -11,16 +11,14 @@ pub fn mozilla_timestamp(timestamp: u64) -> Option<u64> {
   unix_timestamp(timestamp)
 }
 
-#[cfg(target_os = "windows")]
-pub fn internet_explorer_timestamp(timestamp: u64) -> Option<u64> {
-  if timestamp <= 0 {
+fn unix_timestamp(timestamp: u64) -> Option<u64> {
+  if timestamp == 0 {
     return None;
   }
-  let mut timestamp = timestamp - 116_444_736_000_000_000;
-  timestamp /= 10_000_000;
-  unix_timestamp(timestamp)
+  Some(timestamp)
 }
 
+#[cfg(target_os = "macos")]
 pub fn safari_timestamp(timestamp: u64) -> Option<u64> {
   if timestamp == 0 {
     return None;
@@ -30,10 +28,12 @@ pub fn safari_timestamp(timestamp: u64) -> Option<u64> {
   let unix_timestamp = unix_timestamp / 1_000_000_000;
   Some(unix_timestamp)
 }
-
-fn unix_timestamp(timestamp: u64) -> Option<u64> {
-  if timestamp == 0 {
+#[cfg(target_os = "windows")]
+pub fn internet_explorer_timestamp(timestamp: u64) -> Option<u64> {
+  if timestamp <= 0 {
     return None;
   }
-  Some(timestamp)
+  let mut timestamp = timestamp - 116_444_736_000_000_000;
+  timestamp /= 10_000_000;
+  unix_timestamp(timestamp)
 }
