+++ rookie-rs/src/common/mod.rs
@@ -3,9 +3,3 @@ pub mod enums;
 pub mod paths;
 pub mod sqlite;
 pub mod utils;
-
-#[cfg(unix)]
-pub mod secrets;
-
-#[cfg(target_os = "windows")]
-pub mod winapi;
