+++ rookie-rs/src/common/utils.rs
@@ -8,12 +8,3 @@ pub fn some_domain_in_host(domains: Option<Vec<&str>>, host: &str) -> bool {
   }
   false
 }
-
-#[cfg(target_os = "linux")]
-pub fn capitalize(s: &str) -> String {
-  let mut c = s.chars();
-  match c.next() {
-    None => String::new(),
-    Some(f) => f.to_uppercase().collect::<String>() + c.as_str(),
-  }
-}
