+++ rookie-rs/src/linux/mod.rs
@@ -1,0 +1,5 @@
+mod config;
+pub mod path;
+pub mod secrets;
+
+pub const APP_ID: &str = "rookie";
