+++ rookie-rs/src/windows/dpapi.rs
